pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentpermit"
version = "0.1.0"
description = "Open-source permission, audit, and safety layer for AI agents."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "AgentPermit contributors" }]
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[project.scripts]
agentpermit = "agentpermit.cli:app"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
